Move some srv6 functions to srv6_helper.py (#18892)
